In f_sphere.c: Change on the order of "gradient" and the search point "x" in the function "f_sphere_evaluate_gradient".
In suite_consprob2015.c: Changes in the function "suite_consprob2015" in order to follow the recent modifications in "coco_internal.h". More specifically, use of the attributes "suite_dep_index", "suite_dep_function_id" and "suite_dep_instance_id". Also, a little modification has been made on the type of the variable "dimension" from long to int, as it is now in "suite_bbob2009.c".