bbob logger now always logs the number of constraint evaluations
which in the unconstrained case is always zero. The logger
code has also become shorter and more comprehensible.

The code is based on commit dc09472f by @asmaatamna which however has been
lost in the merge process. Should address issues #1518 #1444 #1160 #1367

The constraints evaluation column may or may not break the backwards
compatibility to post-processing routines.

	modified:   code-experiments/src/logger_bbob.c