simplified non-anytime example, trying to minimize performance assessment errors,
and semibug-fix from restarts and using final target hit.

For a budgeted algorithm it seems somewhat meaningless to do automated restarts.
They seem to defeat the purpose of a non-anytime assessment. Hence the restarts have
been removed with the benefit of code simplification also concering the `delivered`
method, which doesn't need to make a case distinction.

	modified:   code-experiments/build/python/example_experiment_non_anytime.py