a data archive class addressing issue #1533
yet somewhat a draft but hopefully usable and useful.

cocopp.data_archive is the class instantiation through which we
can now access all archived bbob data. The bbob-biobj and bbob-noisy
data have yet to be added to the list of known data (and no more
than that). This can be done manually, or with a small script
which could become part of the class.

	modified:   code-postprocessing/cocopp/__init__.py
	modified:   code-postprocessing/cocopp/findfiles.py