Updating comments to the code
Started to play around with the pycma package and cleaned a bit the src code by writing functions for allocating the distributon theta array for each distribution

Fixing issue with indexing of problems, function and other stuff

Adding best_decision_vector to the python interface

Cleaning up before PR, Removing unnecessary files

Final Commit

This should be the "final" commit, in the sense that it is the first
commit where the results obtained are exactly the same as the one for
the legacy code. Before I forgot to implement the boundary handling
method that were adding a penalty on the decision vector.

Hopefully this is the right time, but I guess we'll have to see the results of the tests done at commit time

Removing printf statemets after debugging

Passing structures as void *