[cocoex/core] Write version string to C file.
Previously we stored the version string in a statically linked variable,
declared in the coco.h header. While this works fine on Linux and macOS,
it leads to link errors on Windows when the Rust bindings (coco-sys) try
to access the coco_version variable. This seems to stem from MSVC
compiling the source as C++.