[build] Allow fabrication outside git repository
Currently fabricate expects to be run from inside a git checkout so
that it can derive the version number from the git tags. This breaks
building cocopp / cocoex from a ZIP file download from GitHub.

Instead of erring out, this commit uses a default version number of
'99.0.1' for the build.

Fixes #2259